+++ src/Concerto/PanelBundle/Resources/public/angularjs/app/concerto_panel/js/controllers/content_controller.js
@@ -15,6 +15,14 @@ function ContentController($scope, $state, $location) {
         name: "administration"
     }];
 
+    $scope.gridFilters = {
+        tests: [],
+        templates: [],
+        tables: [],
+        wizards: [],
+        users: []
+    };
+
     $scope.tab = {
         activeIndex: -1
     };